requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.2.0
numpy>=1.26.0
PyPDF2>=3.0.0
PyYAML>=6.0.0
rich>=13.7.0
streamlit>=1.31.0
typer>=0.12.0
pytest>=8.0.0
rouge-score>=0.1.2
